backend done

+++ resolvers/Mutation/usuario.js
@@ -15,7 +15,8 @@ const mutations = {
         })
     },
 
-    async novoUsuario(_, { dados }) {
+    async novoUsuario(_, { dados }, ctx) {
+        ctx && ctx.validarAdmin()
         try {
             const idsPerfis = []
             if (!dados.perfis || !dados.perfis.length) {
@@ -50,7 +51,8 @@ const mutations = {
             throw new Error(e.sqlMessage)
         }
     },
-    async excluirUsuario(_, args) {
+    async excluirUsuario(_, args, ctx) {
+        ctx && ctx.validarAdmin()
         try {
             const usuario = await obterUsuario(_, args)
             if (usuario) {
@@ -66,12 +68,13 @@ const mutations = {
         }
 
     },
-    async alterarUsuario(_, { filtro, dados }) {
+    async alterarUsuario(_, { filtro, dados }, ctx) {
+        ctx && ctx.validarUsuarioFiltro(filtro)
         try {
             const usuario = await obterUsuario(_, { filtro })
             if (usuario) {
                 const { id } = usuario
-                if (dados.perfis) {
+                if (ctx.admin && dados.perfis) {
                     await db('usuarios_perfis')
                         .where({ usuario_id: id }).delete()
